[listview] Replaced listview by recycledlistview
There was a lag of 1 second on the ListView when transition between
selections. That was totally unacceptable.  Swapped to recycled list
view and cached the selected and not selected.
Tried to have an animation on the transition but was not working very
well. Wil try again later.

Header has to be reimplemented.